Cleaning up the poller top-level script.
