+++ src/common/helpers/poller.js
@@ -60,7 +60,8 @@ export const hasRepoChanged = async (repositoryUrl, last_updated_at, token) => {
 
 // Whether a given snap (GitHub) repository has changed since 'last_updated_at'.
 // Consider changes in the repository itself any of the (GitHub) parts source.
-export const checkSnapRepository = async (owner, name, last_updated_at, token) => {
+export const checkSnapRepository = async (owner, name, last_updated_at) => {
+  const token = conf.get('GITHUB_AUTH_CLIENT_TOKEN');
   const repo_url = getGitHubRepoUrl(owner, name);
   if (await hasRepoChanged(repo_url, last_updated_at, token)) {
     return true;
@@ -80,4 +81,34 @@ export const checkSnapRepository = async (owner, name, last_updated_at, token) =
     }
   }
   return false;
+};
+
+
+// XXX: meh no ES6 import support, great library :-/
+let AsyncLock = require('async-lock');
+let processRepoLock = new AsyncLock();
+
+
+// Process a given Repository (DB) model. Check for changes using
+// `checkSnapRepository` and if changed request a LP snap build and mark
+// it as 'updated'.
+export const processRepository = (repo) => {
+  processRepoLock.acquire('PROCESS-REPO-SYNC', async () => {
+    const owner = repo.get('owner');
+    const name = repo.get('name');
+    const last_updated_at = repo.get('updated_at');
+
+    logger.info(`${owner}/${name}: Polling ...`);
+    try {
+      if (await checkSnapRepository(owner, name, last_updated_at)) {
+        logger.info(`${owner}/${name}: NEEDSBUILD`);
+        // XXX request build and reset repo.updated_at.
+      } else {
+        logger.info(`${owner}/${name}: UNCHANGED`);
+      }
+    } catch (e) {
+      logger.error(`${owner}/${name}: FAILED (${e.message})`);
+    }
+    logger.info('==========');
+  });
 };
